Create a separate store for trending podcasts
