+++ src/lib/entities/podcast/index.ts
@@ -1,4 +1,5 @@
 export { default as PodcastTile } from './ui/podcast-tile.svelte';
 export { default as PodcastDisplay } from './ui/podcast-display.svelte';
 export { podcastList, podcastGet } from './api/requests';
-export { podcasts, getPodcastByID, addPodcastToStore } from './model/store';
+export { podcasts, getPodcastByID, addPodcastToStore } from './model/podcast-store';
+export { trendingPodcasts } from './model/trending-store';
