+++ src/lib/entities/podcast/model/trending-store.ts
@@ -1,0 +1,4 @@
+import { writable } from 'svelte/store';
+import type { Podcast } from '$lib/shared/api';
+
+export const trendingPodcasts = writable<Podcast[]>([]);
